EDC-30
EDC-30 | Angga | Bookmark

+++ database/migrations/2025_04_09_151346_create_comments_table.php
@@ -11,22 +11,23 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('comments', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('matkul_id');
-            $table->unsignedBigInteger('user_id');
-            $table->text('comment');
-            $table->timestamps();
-    
-            // Foreign Key Constraints
-            $table->foreign('matkul_id')->references('id')->on('mata_kuliahs')->onDelete('cascade');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+        if (!Schema::hasTable('comments')) {
+            Schema::create('comments', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('matkul_id');
+                $table->unsignedBigInteger('user_id');
+                $table->text('comment');
+                $table->timestamps();
+
+                // Foreign Key Constraints
+                $table->foreign('matkul_id')->references('id')->on('mata_kuliahs')->onDelete('cascade');
+                $table->foreign('user_id')->references('id')->on('penggunas')->onDelete('cascade');
+            });
+        }
     }
-    
+
     public function down()
     {
         Schema::dropIfExists('comments');
     }
-    
 };
